Moved code from Claw.java to ClawJoint.java
